refactor actor armor, monster armor ablates
